requirements.txt
moviepy==1.0.3
srt==3.5.3
demucs==4.0.1
deep_translator==1.11.4
openai-whisper==20240930